services/backend/requirements.txt
# Offline install using vendor_wheels:
# pip install --no-index --find-links=../vendor_wheels -r requirements.txt
fastapi
uvicorn
pydantic
qrcode[pil]
requests
paho-mqtt
openpyxl
python-multipart
pandas
jinja2
omise
line-bot-sdk
schedule
